main

+++ src/main/java/com/capstone/ai_painter_backen/service/UserService.java
@@ -52,9 +52,12 @@ public class UserService {
     public void deleteUser(UserDto.UserDeleteDto deleteDto){
         UserEntity userEntity = userRepository.findById(deleteDto.getUserId())
                 .orElseThrow(()-> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
-        if(!userEntity.getProfileImage().equals(DEFAULT_PROFILE)){
+
+
+         if(userEntity.getProfileImage() != null &&
+                 !userEntity.getProfileImage().equals(DEFAULT_PROFILE)){
             s3FileService.deleteMultiFile(userEntity.getProfileImage());
-        }
+        }//todo 수정 완료!!!
         userRepository.deleteById(deleteDto.getUserId());
         log.info("{}: User 삭제됨 !", deleteDto.getUserId());
     }
@@ -88,7 +91,7 @@ public class UserService {
             patchDto.setNickname(userEntity.getUserRealName());
         }
         if (patchDto.getPassword() == null) {
-            patchDto.setNickname(userEntity.getPassword());
+            patchDto.setPassword(userEntity.getPassword());
         }
 
         userEntity.update(patchDto, s3ImageInfo);
